kc_token_cache/requirements.txt
cachetools>=5.0.0
PyYAML>=6.0
httpx>=0.24.0
fastapi>=0.95.0
uvicorn>=0.22.0